requirements.txt
discord.py>=2.3.2
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.4
flask>=2.3.2
waitress>=2.1.2
deep-translator>=1.11.4
langdetect>=1.0.9

